ExecTimer class as its own file
